FS-5062 - Reorganising import_config tests by combining and renaming the two existing test modules
